+++ tests/react/armor-control.test.js
@@ -51,4 +51,56 @@ describe('ArmorControl', function() {
 
         expect(spy).toHaveBeenCalledWith(null);
     });
+
+    it("accounts for explicit damage reduction from armor quality", async function () {
+        // AVERAGE over leth reduction types * 2 /3
+        // =-ROUNDUP(POWER(2/3*AVERAGE(AY36:BB36);2);0)
+        // => -9 overall leth reduction results in 36 DR
+        render(<ArmorControl campaign={2} armor={factories.armorFactory({
+            base: factories.armorTemplateFactory({'armor_t_p': -9, 'armor_t_dr': -14}),
+            quality: factories.armorQualityFactory({'armor_dr': -3})
+        })} />)
+        expect(screen.getByLabelText("Armor T DR").textContent).toEqual("-17")
+    });
+
+    it("accounts for automatic damage reduction from armor quality", async function () {
+        // AVERAGE over leth reduction types * 2 /3
+        // =-ROUNDUP(POWER(2/3*AVERAGE(AY36:BB36);2);0)
+        // => -9 overall leth reduction results in 36 DR
+        render(<ArmorControl campaign={2}
+                             armor={factories.armorFactory({
+            base: factories.armorTemplateFactory({'armor_t_p': -9, 'armor_t_s': -9, 'armor_t_b': -9, 'armor_t_r': -9, 'armor_t_dr': -36}),
+            quality: factories.armorQualityFactory({'armor_p': -0.5, 'armor_s': -0.5, 'armor_b': -0.5, 'armor_r': -0.5, 'armor_dr': 0})
+        })}
+                             helm={factories.armorFactory({
+                base: factories.armorTemplateFactory({'armor_h_p': -9, 'armor_h_s': -9, 'armor_h_b': -9, 'armor_h_r': -9, 'armor_h_dr': -36}),
+            quality: factories.armorQualityFactory({'armor_p': -0.5, 'armor_s': -0.5, 'armor_b': -0.5, 'armor_r': -0.5, 'armor_dr': 0})
+        })}
+        />)
+        expect(screen.getByLabelText("Armor T DR").textContent).toEqual("-40")
+        expect(screen.getByLabelText("Armor H DR").textContent).toEqual("-40")
+
+        expect(screen.getByLabelText("Armor LL DR").textContent).toEqual("0")
+        expect(screen.getByLabelText("Armor LL P").textContent).toEqual("0")
+    });
+
+    it("skips automatic damage reduction from armor quality when quality does not affect location", async function () {
+        const user = userEvent.setup()
+
+        render(<ArmorControl campaign={2} armor={factories.armorFactory({ name: "FooArmor",
+            base: factories.armorTemplateFactory({'armor_t_p': -9.2, 'armor_t_s': -9.2, 'armor_t_b': -9.2, 'armor_t_r': -9.2, 'armor_t_dr': -36}),
+            quality: factories.armorQualityFactory({'armor_p': 0, 'armor_s': 0, 'armor_b': 0, 'armor_r': 0, 'armor_dr': 0})
+        })} />)
+        const torsoLocation = screen.getByLabelText("Armor T DR");
+        expect(torsoLocation.textContent).toEqual("-36")
+
+        await user.click(torsoLocation)
+
+        const breakdown = screen.getByLabelText("Armor T DR breakdown")
+
+        expect(breakdown).toBeVisible()
+
+        // Find the breakdown row for the armor and check that it has the correct value.
+        expect(within(within(breakdown).getByText("FooArmor").closest("tr")).getAllByRole("cell")[1].textContent).toEqual("-36")
+    });
 });